+++ test/unit/routing/url_test.dart
@@ -4,11 +4,36 @@ import 'package:test/test.dart';
 void main() {
   test('uri generation', () {
     final d = StandardUriDesign.pathOnly;
+    expect(d.base.path, equals('/'));
     expect(d.collection('books').toString(), '/books');
     expect(d.resource('books', '42').toString(), '/books/42');
     expect(d.related('books', '42', 'author').toString(), '/books/42/author');
     expect(d.relationship('books', '42', 'author').toString(),
         '/books/42/relationships/author');
+  });
+
+  test('uri generation with base, trailing slash', () {
+    final d = StandardUriDesign(Uri.parse('https://example.com/api/'));
+    expect(d.base.path, equals('/api/'));
+    expect(d.collection('books').toString(), 'https://example.com/api/books');
+    expect(d.resource('books', '42').toString(),
+        'https://example.com/api/books/42');
+    expect(d.related('books', '42', 'author').toString(),
+        'https://example.com/api/books/42/author');
+    expect(d.relationship('books', '42', 'author').toString(),
+        'https://example.com/api/books/42/relationships/author');
+  });
+
+  test('uri generation with base, no trailing slash', () {
+    final d = StandardUriDesign(Uri.parse('https://example.com/api'));
+    expect(d.base.path, equals('/api/'));
+    expect(d.collection('books').toString(), 'https://example.com/api/books');
+    expect(d.resource('books', '42').toString(),
+        'https://example.com/api/books/42');
+    expect(d.related('books', '42', 'author').toString(),
+        'https://example.com/api/books/42/author');
+    expect(d.relationship('books', '42', 'author').toString(),
+        'https://example.com/api/books/42/relationships/author');
   });
 
   test('Authority is retained if exists in base', () {
@@ -79,6 +104,40 @@ void main() {
 
     test('Authority and path', () {
       final d = StandardUriDesign(Uri.parse('https://example.com:8080/api'));
+      expect(d.matchTarget(Uri.parse('https://example.com:8080/api/books')),
+          isA<Target>().having((it) => it.type, 'type', equals('books')));
+      expect(
+          d.matchTarget(Uri.parse('https://example.com:8080/api/books/42')),
+          isA<ResourceTarget>()
+              .having((it) => it.type, 'type', equals('books'))
+              .having((it) => it.id, 'id', equals('42')));
+      expect(
+          d.matchTarget(
+              Uri.parse('https://example.com:8080/api/books/42/authors')),
+          isA<RelatedTarget>()
+              .having((it) => it.type, 'type', equals('books'))
+              .having((it) => it.id, 'id', equals('42'))
+              .having(
+                  (it) => it.relationship, 'relationship', equals('authors')));
+      expect(
+          d.matchTarget(Uri.parse(
+              'https://example.com:8080/api/books/42/relationships/authors')),
+          isA<RelationshipTarget>()
+              .having((it) => it.type, 'type', equals('books'))
+              .having((it) => it.id, 'id', equals('42'))
+              .having(
+                  (it) => it.relationship, 'relationship', equals('authors')));
+
+      expect(
+          d.matchTarget(Uri.parse('https://example.com:8080/a/b/c/d')), isNull);
+      expect(d.matchTarget(Uri.parse('http://example.com:8080/books')), isNull);
+      expect(d.matchTarget(Uri.parse('https://foo.net:8080/books')), isNull);
+      expect(d.matchTarget(Uri.parse('https://example.com:8080/foo/books')),
+          isNull);
+    });
+
+    test('Authority and path, trailing slash', () {
+      final d = StandardUriDesign(Uri.parse('https://example.com:8080/api/'));
       expect(d.matchTarget(Uri.parse('https://example.com:8080/api/books')),
           isA<Target>().having((it) => it.type, 'type', equals('books')));
       expect(